uadefs: add sound data cache [PERFECTIVE]
The sound data cache is called a stash. It contains synthesized sample
data from previously played songs. This was implemented to optimize XMMS
file scan. XMMS passes a given WAV file for several plugins before the
WAV plugin. Each plugin usually reads some data that
causes uade123 to be spawned. It is wasteful to spawn uade123 for all of
them. The amount of data that is read by the plugins is usually small,
so we cache it. We synthesize the data once, and pass the same data for
all plugins.

This more than doubled the speed of "xmms -e /mod/dir" on my system.