configure: use $CC --version, not -v
The -v compiler option is actually the verbose option switch, not
explicitly version. In some circumstances it will have unexpected
effects; for instance, depending on the commands passed, it's possible
for `$CC -v` to exit non-zero, which will cause the configure script
to think the compiler doesn't work.

An example of this is the Homebrew package manager's "superenv" build
environment, which wraps the compilers and adds extra unrequested flags
to ensure a consistent build environment. One of these happens to cause
`$CC -v` to exit non-zero, without affecting the behaviour of
`$CC --version`.